requirements.txt
django-widget-tweaks
#widget_tweaks
django-ckeditor
#ckeditor
Pillow
#Django(LTS)
Django==2.2.5
#strictly separate the settings parameters
python-decouple
#safe db url,credentials
dj-database-url
